[add] SOを作り直し
