+++ Assets/Scripts/Runtime/5.InfraStructure/OutGame/Scenario/ScenarioSettingsAsset.asset
@@ -9,11 +9,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 6cbc8065f2671a841b110549a652fc11, type: 3}
   m_Name: ScenarioSettingsAsset
-  m_EditorClassIdentifier: KillChord.Structure:KillChord.Runtime.InfraStructure:ScenarioSettingsAsset
-  _normalTextCharIntervalSec: 0.2
-  _fastForwardTextCharIntervalSec: 0.02
+  m_EditorClassIdentifier: KillChord.Structure::KillChord.Runtime.InfraStructure.ScenarioSettingsAsset
+  _normalTextCharIntervalSec: 0.4
+  _fastForwardTextCharIntervalSec: 0.2
   _pausePollIntervalSec: 0.05
   _closeDelayAfterCompleteSec: 3
   _skipClosesImmediately: 1
