+++ Assets/Scripts/Runtime/5.InfraStructure/OutGame/Scenario/ScenarioSettingsAsset.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6cbc8065f2671a841b110549a652fc11
